create new instances of the classes
When working with classes, we need to create a new instances of the classes we use. The creation of the instance gets assigned to a local variable, which the class automatically invokes __init__() for the newly created class instance. So the added changes helps define certain variables that updates the color palette and toggling on the LED matrix.